toolkit/content/tests/widgets/test_editor_currentURI.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin"
                 type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Editor currentURI Tests" onload="runTest();">
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <body xmlns="http://www.w3.org/1999/xhtml">
  <p/>
  <editor xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
          id="editor"
          type="content"
          editortype="html"
          style="width: 400px; height: 100px;"/>
  <p/>
  <pre id="test">
  </pre>
  </body>
  <script class="testbody" type="application/javascript">
  <![CDATA[

  SimpleTest.waitForExplicitFinish();

  function runTest() {
    var editor = document.getElementById("editor");
    // Check that currentURI is a property of editor.
    var result = "currentURI" in editor;
    is(result, true, "currentURI is a property of editor");
    is(editor.currentURI.spec, "about:blank", "currentURI.spec is about:blank");
    SimpleTest.finish();
  }
]]>
</script>
</window>
